requirements.txt
wget
nemo_toolkit[asr]==2.0.0rc0
nltk
faster-whisper>=1.1.0
yt-dlp
python-docx
git+https://github.com/MahmoudAshraf97/demucs.git
git+https://github.com/oliverguhr/deepmultilingualpunctuation.git
git+https://github.com/MahmoudAshraf97/ctc-forced-aligner.git

